requirements.txt
# Core dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# AI providers
google-generativeai==0.8.3

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
chardet==5.2.0

# Data visualization
plotly==5.18.0

# HTTP requests
requests==2.31.0

# XML parsing for arXiv
lxml==4.9.3

# Utilities
dataclasses-json==0.6.3

# Optional for enhanced functionality
pandas==2.1.4
numpy==1.26.2

# Development dependencies (optional)
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# Logging
python-json-logger==2.0.7